更改

+++ src/main/java/com/example/tomatomall/repository/AccountRepository.java
@@ -3,7 +3,7 @@ package com.example.tomatomall.repository;
 import com.example.tomatomall.po.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<Account,Long>
+public interface AccountRepository extends JpaRepository<Account,Integer>
 {
     Account findByUsername(String username);
 }
